third_party/WebKit/LayoutTests/css3/device-adapt/viewport-relative-font.html

<!DOCTYPE html>
<html>
<head>
    <title>Font relative lengths in @viewport relative to initial font</title>
    <style>
        html { font-size: 100px; }
    </style>
</head>
<body>
    <script>
        document.body.offsetWidth; // Trigger style resolving
    </script>
    <style>
        @viewport {
            width: 10em;
            height: 10rem;
        }
    </style>
    <script>
        if (window.testRunner) {
            testRunner.dumpAsText();
            document.body.offsetWidth; // Trigger style resolving
            alert(internals.viewportAsText(document, 1, 320, 352));
        }
    </script>
</body>
</html>
