third_party/WebKit/LayoutTests/css3/device-adapt/viewport-relative-font-expected.txt

ALERT: viewport size 160x160 scale 2 with limits [2, -1] and userScalable true
